+++ lib/builder.dart
@@ -3,7 +3,6 @@ library flutter_riverpod_generator.builder;
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 
-import 'src/floor_dao_riverpod_generator.dart';
 import 'src/floor_database_riverpod_generator.dart';
 
 /// Entry point for build_runner
@@ -12,7 +11,6 @@ import 'src/floor_database_riverpod_generator.dart';
 Builder floorRiverpodProviderBuilder(BuilderOptions options) =>
     SharedPartBuilder(
       [
-        FloorDaoRiverpodGenerator(),
         FloorDatabaseRiverpodGenerator(),
       ],
       'floorRiverpod',