+++ lib/src/annotations.dart
@@ -1,4 +1,0 @@
-class DaoProvider {
-  final Type databaseType;
-  const DaoProvider(this.databaseType);
-}