+++ lib/src/floor_database_riverpod_generator.dart
@@ -2,19 +2,20 @@ import 'dart:async';
 
 import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
-import 'package:floor_annotation/floor_annotation.dart';
+import 'package:floor_annotation/floor_annotation.dart' as floor;
+import 'package:floor_riverpod_generator/src/floor_dao_riverpod_generator.dart';
 import 'package:source_gen/source_gen.dart';
 
-import 'annotations.dart';
 import 'cfg.dart';
 
-class FloorDatabaseRiverpodGenerator extends GeneratorForAnnotation<Database> {
+class FloorDatabaseRiverpodGenerator
+    extends GeneratorForAnnotation<floor.Database> {
   @override
   FutureOr<String> generateForAnnotatedElement(
     Element element,
     ConstantReader annotation,
     BuildStep buildStep,
-  ) {
+  ) async {
     if (element is! ClassElement) return '';
     final T = element.name;
 
@@ -29,9 +30,9 @@ class FloorDatabaseRiverpodGenerator extends GeneratorForAnnotation<Database> {
       '  //Some DAOs need direct access to the database',
     ];
 
-    for (final daoElement in daos) {
-      //Call other generator?
+    final daoProviders = <String>[];
 
+    for (final daoElement in daos) {
       final dao = daoElement.type.element as ClassElement;
       final daoFieldName = daoElement.name;
       final dbPtr = dao.fields.where((field) =>
@@ -41,17 +42,28 @@ class FloorDatabaseRiverpodGenerator extends GeneratorForAnnotation<Database> {
         final targetName = dbPtr.first.name;
         out.add('db.$daoFieldName.$targetName = db.database;');
       }
+
+      //Call other generator?
+      daoProviders.add(await _daoGen.directGenerateForDao(
+        dao,
+        element,
+        dbProvider,
+      ));
     }
 
     out.addAll([
       '  return db;',
       '});',
+      '',
+      daoProviders.join('\n\n'),
     ]);
 
     return out.join('\n');
   }
 
-  final _daoType = const TypeChecker.fromRuntime(DaoProvider);
+  final _daoGen = FloorDaoRiverpodGenerator();
+
+  final _daoType = TypeChecker.fromRuntime(floor.dao.runtimeType);
   // final _dbExecType = const TypeChecker.fromRuntime(sqflite.DatabaseExecutor);
 
   bool _isDao(final Element? e) {