+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## 1.0.0
+## 0.3.0
+- Transition to entirely use Floor annotations
+- Fix: critical build error to run since annotations in package
+- Breaking: All generated providers under database
+
+## 0.2.0
+
+- Add provider for database
+
+## 0.1.0
 
 - Initial version.